quickfixes for compiler warnings

+++ test/wh_test_clientserver.c
@@ -512,7 +512,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != 0) {
         WH_ERROR_PRINT(
             "Failed to add non-exportable NVM object: ret=%d, out_rc=%d\n", ret,
-            out_rc);
+            (int)out_rc);
         return ret != 0 ? ret : out_rc;
     }
 
@@ -523,7 +523,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != WH_ERROR_ACCESS) {
         WH_ERROR_PRINT("Non-exportable NVM object was read unexpectedly: "
                        "ret=%d, out_rc=%d\n",
-                       ret, out_rc);
+                       ret, (int)out_rc);
         return -1;
     }
 
@@ -543,7 +543,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != 0) {
         WH_ERROR_PRINT(
             "Failed to add exportable NVM object: ret=%d, out_rc=%d\n", ret,
-            out_rc);
+            (int)out_rc);
         return ret != 0 ? ret : out_rc;
     }
 
@@ -556,7 +556,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != 0) {
         WH_ERROR_PRINT(
             "Failed to read exportable NVM object: ret=%d, out_rc=%d\n", ret,
-            out_rc);
+            (int)out_rc);
         return ret != 0 ? ret : out_rc;
     }
 
@@ -586,7 +586,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != 0) {
         WH_ERROR_PRINT("Failed to add non-exportable NVM object for DMA: "
                        "ret=%d, out_rc=%d\n",
-                       ret, out_rc);
+                       ret, (int)out_rc);
         return ret != 0 ? ret : out_rc;
     }
 
@@ -598,7 +598,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != WH_ERROR_ACCESS) {
         WH_ERROR_PRINT("Non-exportable NVM object was read via DMA "
                        "unexpectedly: ret=%d, out_rc=%d\n",
-                       ret, out_rc);
+                       ret, (int)out_rc);
         return -1;
     }
 
@@ -617,7 +617,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != 0) {
         WH_ERROR_PRINT(
             "Failed to add exportable NVM object for DMA: ret=%d, out_rc=%d\n",
-            ret, out_rc);
+            ret, (int)out_rc);
         return ret != 0 ? ret : out_rc;
     }
 
@@ -629,7 +629,7 @@ static int _testNonExportableNvmAccess(whClientContext* client)
     if (ret != 0 || out_rc != 0) {
         WH_ERROR_PRINT(
             "Failed to read exportable NVM object via DMA: ret=%d, out_rc=%d\n",
-            ret, out_rc);
+            ret, (int)out_rc);
         return ret != 0 ? ret : out_rc;
     }
